feat: add a data.js,and all of data import this file to load data
