+++ src/views/data/data.js
@@ -1,0 +1,58 @@
+const talklist = [
+  {
+    id: '1',
+    picture: 'https://tdesign.gtimg.com/mobile/demos/avatar2.png',
+    count: '0',
+    name: 'Pite',
+    message: [
+      { id: '1', tag: 'other', value: '那明天准时见哦' },
+      { id: '2', tag: 'me', value: '好的，我会记得的' },
+      { id: '3', tag: 'me', value: '在吗？' },
+      { id: '4', tag: 'time', value: '今天 10:50' },
+      {
+        id: '5',
+        tag: 'other',
+        value: '有个问题想咨询一下，关于Tdesign组件库如何更好的使用',
+      },
+      { id: '6', tag: 'me', value: '你请问' },
+    ],
+  },
+  {
+    id: '2',
+    picture: 'https://tdesign.gtimg.com/mobile/demos/avatar2.png',
+    count: '2',
+    name: 'Bob',
+    message: [
+      { id: '1', tag: 'other', value: '那明天准时见哦' },
+      { id: '2', tag: 'me', value: '好的，我会记得的' },
+      { id: '3', tag: 'me', value: '在吗？' },
+      { id: '4', tag: 'time', value: '今天 10:50' },
+      {
+        id: '5',
+        tag: 'other',
+        value: '有个问题想咨询一下，关于Tdesign组件库如何更好的使用',
+      },
+      { id: '6', tag: 'me', value: '你请问' },
+    ],
+  },
+  {
+    id: '3',
+    picture: 'https://tdesign.gtimg.com/mobile/demos/avatar2.png',
+    count: '0',
+    name: 'Alice',
+    message: [
+      { id: '1', tag: 'other', value: '那明天准时见哦' },
+      { id: '2', tag: 'me', value: '好的，我会记得的' },
+      { id: '3', tag: 'me', value: '在吗？' },
+      { id: '4', tag: 'time', value: '今天 10:50' },
+      {
+        id: '5',
+        tag: 'other',
+        value: '有个问题想咨询一下，关于Tdesign组件库如何更好的使用',
+      },
+      { id: '6', tag: 'me', value: '你请问' },
+    ],
+  },
+]
+export { talklist }
+export default talklist
